fix compute type detection logic to allow user-defined settings

+++ src/config/settings.py
@@ -50,7 +50,7 @@ PROCESSING_CONFIG = {
     'max_workers': 2,
     'default_model': 'large-v3',
     'default_language': 'ru',
-    'default_compute_type': 'float32',
+    'default_compute_type': 'float16',
     'default_batch_size': 4,
     'default_diarize': True
 }
